+++ routes/leaderRouter.js
@@ -19,27 +19,42 @@ leaderRouter.route('/')
         .catch((err) => next(err));
     })
     .post(authenticate.veryUser, (req, res, next) => {
-            Leaders.create(req.body)
-            .then((leader)=>{
-                    console.log('Created Leader ', leader);
-                    res.statusCode=200;
-                    res.setHeader('Content-Type','application/json');
-                    res.json(leader);
-            },(err) => next(err) )
-            .catch((err) => next(err));
+        if(authenticate.verifyAdmin(req.user)){
+                Leaders.create(req.body)
+                .then((leader)=>{
+                        console.log('Created Leader ', leader);
+                        res.statusCode=200;
+                        res.setHeader('Content-Type','application/json');
+                        res.json(leader);
+                },(err) => next(err) )
+                .catch((err) => next(err));
+        }
+        else{
+                err = new Error('Only admins can do that!!');
+                err.statusCode = 403;
+                next(err);
+        } 
+            
     })
     .put(authenticate.veryUser, (req, res, next) => {
             res.statusCode = 403;
             res.end('PUT operation not supported on /leaders');
     })
     .delete(authenticate.veryUser, (req, res, next) => {
-            Leaders.deleteMany({})
+        if(authenticate.verifyAdmin(req.user)){
+        Leaders.deleteMany({})
             .then((resp)=>{
                     res.statusCode=200;
                     res.setHeader('Content-Type','application/json');
                     res.json(resp);
             },(err) => next(err) )
             .catch((err) => next(err));
+        }
+        else{
+                err = new Error('Only admins can do that!!');
+                err.statusCode = 403;
+                next(err);
+        }    
     });
     
     
@@ -57,17 +72,25 @@ leaderRouter.route('/:leaderId')
         res.end('POST operation not supported on /leaders/'+ req.params.leaderId);
 })
 .put(authenticate.veryUser, (req, res, next) => {
-        Leaders.findByIdAndUpdate(req.params.leaderId,{
-                    $set:req.body
-            },{ new:true })
-            .then((leader)=>{
-                    res.statusCode=200;
-                    res.setHeader('Content-Type','application/json');
-                    res.json(leader);
-            },(err) => next(err) )
-            .catch((err) => next(err));
+        if(authenticate.verifyAdmin(req.user)){
+                Leaders.findByIdAndUpdate(req.params.leaderId,{
+                        $set:req.body
+                },{ new:true })
+                .then((leader)=>{
+                        res.statusCode=200;
+                        res.setHeader('Content-Type','application/json');
+                        res.json(leader);
+                },(err) => next(err) )
+                .catch((err) => next(err));
+        }
+        else{
+                err = new Error('Only admins can do that!!');
+                err.statusCode = 403;
+                next(err);
+        } 
 })
 .delete(authenticate.veryUser, (req, res, next) => {
+        if(authenticate.verifyAdmin(req.user)){
         Leaders.findByIdAndDelete(req.params.leaderId)
             .then((resp)=>{
                     res.statusCode=200;
@@ -75,6 +98,12 @@ leaderRouter.route('/:leaderId')
                     res.json(resp);
             },(err) => next(err) )
             .catch((err) => next(err));
+        }
+        else{
+                err = new Error('Only admins can do that!!');
+                err.statusCode = 403;
+                next(err);
+        } 
 });
       
 module.exports = leaderRouter;
